Added a check in the testbench on the result.
We XOR the START_PATTERN in the testbench as a check. This is then XORed
with EVEN_PARITY_BIT to invert the answer if we're set up for outputting
an even parity bit.

Essentially we verify that serialization of the data word is working.

Adding of debug statements in indicate answer.