Set up tasks. Changed test to run through loop of random values.
Setup an init() task for initialization of clk, rst.
Setup a re_init() task for re-init of values for when restarting test.
Setup a reset_pulse_gen() task, which when called generates a reset
pulse.
Changed test so that data_words are randomly generated and passed into
the data_gen module at runtime. This allows for the test to rerun
multiple times now.
Set up an overall failure flag that is set if any of the tests fails.